feat: move form into redux state
