+++ src/store/form/FormInterface.ts
@@ -1,0 +1,12 @@
+export interface FormInterface {
+    text: string,
+    date: string,
+    select: string,
+    checkbox: boolean,
+    switcher: string,
+    file: FileList
+}
+
+export interface FormState {
+    form: FormInterface[]
+}
