+++ src/store/store.ts
@@ -1,10 +1,16 @@
-import { configureStore } from '@reduxjs/toolkit';
-import photoReducer from './news/flickrSlice';
+import {configureStore} from '@reduxjs/toolkit';
+import photoReducer from './flickr/flickrSlice';
+import formReducer from './form/formSlice'
 
 export const store = configureStore({
-  reducer: {
-    photo: photoReducer,
-  },
+    reducer: {
+        photo: photoReducer,
+        form: formReducer,
+    },
+    middleware: (getDefaultMiddleware) =>
+        getDefaultMiddleware({
+            serializableCheck: false
+        }),
 });
 
 export type AppDispatch = typeof store.dispatch;
